Test refactored.
