+++ pytest.ini
@@ -3,3 +3,4 @@
 DJANGO_SETTINGS_MODULE = judyst_web_service.settings
 # -- recommended but optional:
 python_files = tests.py test_*.py *_tests.py
+markers = critical_test: mark test as critical. These tests must to be checked first.
